+++ package.json
@@ -37,6 +37,7 @@
     "express-session": "^1.17.2",
     "mysql2": "^2.3.3",
     "passport": "^0.5.2",
+    "passport-google-oauth20": "^2.0.0",
     "passport-local": "^1.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
@@ -52,6 +53,7 @@
     "@types/express": "^4.17.13",
     "@types/jest": "27.0.2",
     "@types/node": "^16.0.0",
+    "@types/passport-google-oauth20": "^2.0.11",
     "@types/passport-local": "^1.0.34",
     "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
